Fix UTF-8 conversion in cljs
This fixes the issue of emojis (and other higher unicode plane glyphs) not being transformed into utf-8 correctly, and hence giving the wrong hash.
For existing users, there is a backwards compatibility escape hatch `*use-legacy-utf8-conversion*` that can be set to revert to the old algorithm
in case any of those hashes have been stored without the ability to rehash.

The new approach works for any modern version of cljs, but if you use google closure < 2016 it has this bug: https://github.com/google/closure-library/issues/470 . It's not obvious, but it looks fixed as of Dec 2, 2015.